+++ src/main/java/com/zuehlke/pizza/pizzaorderservice/domain/Order.java
@@ -2,7 +2,7 @@ package com.zuehlke.pizza.pizzaorderservice.domain;
 
 import java.util.List;
 
-public record Order(int id, List<OrderItem> orderItems, Channel channel) {
+public record Order(long id, List<OrderItem> orderItems, Channel channel) {
 
    public boolean contains(PizzaType pizzaType) {
       return orderItems.stream().anyMatch(orderItem -> orderItem.pizzaType().equals(pizzaType));
